requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
anyascii==0.3.2
    # via wagtail
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via jsonschema
beautifulsoup4==4.11.2
    # via wagtail
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==40.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-stripe==2.7.3
    # via -r requirements/requirements.in
django==4.1.8
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-cors-headers
    #   django-filter
    #   django-hijack
    #   django-modelcluster
    #   django-otp
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   drf-spectacular
    #   jsonfield
    #   wagtail
django-allauth==0.54.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==9.1
    # via -r requirements/requirements.in
django-cors-headers==3.10.0
    # via -r requirements/requirements.in
django-environ==0.10.0
    # via -r requirements/requirements.in
django-filter==22.1
    # via wagtail
django-hijack==3.3.0
    # via -r requirements/requirements.in
django-modelcluster==6.0
    # via wagtail
django-otp==1.1.6
    # via django-allauth-2fa
django-permissionedforms==0.1
    # via wagtail
django-taggit==3.1.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
django-waffle==3.0.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
drf-spectacular==0.26.2
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonfield==3.1.0
    # via dj-stripe
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.4
    # via celery
l18n==2021.3
    # via wagtail
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.2
    # via wagtail
pillow==9.5.0
    # via wagtail
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
pypng==0.20220715.0
    # via qrcode
pyrsistent==0.19.3
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   celery
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth-2fa
redis==4.5.4
    # via celery
requests==2.28.2
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.19.1
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   html5lib
    #   l18n
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
telepath==0.3
    # via wagtail
typing-extensions==4.5.0
    # via qrcode
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.15
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==4.2.2
    # via -r requirements/requirements.in
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
